clean up logs + clean up inline scripts
